Update rs_comm_Lease_Statement_v2.SSRS.txt
Kathleen Hurley
Nov 20, 2024, 12:18 AM
to Haseeb, me, Sohaib, Muhammad, Jahangir, Abdur

I understand.  I see no difference in the error now, with this new version, unfortunately.

 

I do not have a test environment to provide you.  I am literally uploading the script and running the report; that’s not a test environment, it’s in a live system.  I would like to provide this to you, as we have discussed multiple times; I do not have it to provide to you. I’ve discussed this with Yardi; they do not offer it any more.  Asking for it won’t make it appear. 

 

I need to be able to get this finished, and then we can determine if we have a going-forward option to work together. I can tell this has not been your preferred method of working, and I understand this.  Alternately, we need to determine that we’re not going to be able to finish this report.  At this time, we know that the report at least ran with errors previous to this round of changes; now it does not run at all.  Can we roll back to prior to this round of changes, when the report ran, and make the modifications differently?